+++ .flake8
@@ -1,21 +1,19 @@
 [flake8]
-# See link below for available options
-# https://flake8.pycqa.org/en/latest/user/options.html#options-and-their-descriptions
-# Move this to pyproject.toml as soon as it is supported.
-# See https://gitlab.com/pycqa/flake8/issues/428
+# See https://flake8.pycqa.org/en/latest/user/options.html#options-and-their-descriptions
+# for available options
 
 exclude =
-    .git,
-    .venv,
-    .eggs,
-    .mypy_cache,
-    .pytest_cache,
-    __pycache__,
+    **/__pycache__,
     *.pyc,
-ignore = E203, E501, W503
-max-line-length = 88
+    .venv,
+    build,
+
+# See https://www.flake8rules.com/ for a list of all builtin error codes
+ignore =
+    E203, E501, W503
 per-file-ignores =
-    __init__.py: F401, F403, F405
-    conftest.py: F401, F403, F405
+    __init__.py: F401
+    conftest.py: F401
+
 show_source = True
 statistics = True
